OpenDataset: display errors
We originally had an error parameter that could be queried, as we
thought OpenDataset would be chained by other commands.

However, when we updated OpenDataset to be a standalone command we did
not revise its error handling. We should not have an error message
return, but instead should display the error via the UIService, as
OpenFile.