SaveAsImage: initialize even if dataset is null
We need to handle the case where dataset is null in the initializer,
such that a "failed to initialize" error does not come back to haunt us.
Better will be to fail later with the "a Dataset is required but none
exists" error, since that is precisely what is going on in that case.