SampleImageUploader: show dialog when upload done
This may be a stopgap solution until we have a better TaskService in
place (https://github.com/imagej/imagej/issues/25). But until that time,
it is nice to show a clear dialog box upon either success _or_ failure.

We show the box only in the run() method itself, so that other potential
uses of the API are not infected with calls to the UIService.

Fixes issue #7.