OpenDataset: get rid of weird, broken header thing
There was a MESSAGE visibility "header" field which, I guess(?), was
intended to be displayed in the input harvester dialog as a message
matching the value of the source. However:

1. It didn't actually display any message at all; and

2. The code called source.length() which triggered when the
   _file_ was over 55 characters in length, rather than its _name_ being
   over 55 characters in length. This seems highly likely to be a bug.

Given these problems, and since I can't even tell what the purpose of
that header thing was supposed to be, anyway, I am simply removing it.