SaveAsImage: purge duplicate input
The command wanted an input Dataset as well as an input Display,
onto which it would set the (new) name of the dataset afterwards.

The common case is that the dataset was the active one of a given
ImageDisplay, although that was not explicit in the typing.

Simpler to just ask for an ImageDisplay specifically and save its
active Dataset. This is a high-level command, after all!

Most importantly, this fixes an issue with the command not
working in the legacy UI, because the legacy UI does not support
arbitrary Display inputs -- only ImageDisplay inputs, thanks to
the legacy layer's LegacyImageDisplayService hack.