Migrate usage of deprecated classes
This class now uses the Neighborhood and related classes to define
neighborhoods, and statistic Ops to do the calculations. Therefore the
neighborhood is defined in a slightly different way (the original one
actually uses a bottom-right rectangle). Assignment of
the values from the stats Ops to the new dataset might be able to be
done using Ops.Map.